GET data from server function rather than from api call
