+++ containers/ecr-viewer/src/app/tests/components/Immunizations.test.tsx
@@ -1,18 +1,14 @@
 import { render } from "@testing-library/react";
 import { axe } from "jest-axe";
-import fs from "fs";
-import YAML from "yaml";
 import { Bundle, Immunization } from "fhir/r4";
 import BundleClinicalInfo from "@/app/tests/assets/BundleClinicalInfo.json";
 import { returnImmunizations } from "@/app/view-data/components/common";
+import { loadYamlConfig } from "@/app/api/services/utils";
 
 describe("Immunizations Table", () => {
   let container: HTMLElement;
   beforeAll(() => {
-    const fhirPathFile = fs
-      .readFileSync("./src/app/api/fhirPath.yml", "utf8")
-      .toString();
-    const fhirPathMappings = YAML.parse(fhirPathFile);
+    const fhirPathMappings = loadYamlConfig();
 
     const immunizationsData = [
       {
