+++ containers/ecr-viewer/src/app/tests/fhir-data.test.tsx
@@ -4,9 +4,8 @@
 import fs from "fs";
 import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
 import { mockClient } from "aws-sdk-client-mock";
-import { GET } from "../api/fhir-data/route"; // Adjust the import path to your actual file path
 import { sdkStreamMixin } from "@smithy/util-stream";
-import { NextRequest } from "next/server";
+import { getEcrData } from "@/app/api/services/ecrDataService";
 
 const s3Mock = mockClient(S3Client);
 const stream = sdkStreamMixin(
@@ -39,7 +38,6 @@ describe("GET API Route", () => {
   it("fetches data from S3 and returns a JSON response", async () => {
     const fakeId = "test-id";
     process.env.SOURCE = "s3";
-    const request = new NextRequest(`http://localhost?id=${fakeId}`);
 
     s3Mock
       .on(GetObjectCommand, {
@@ -50,9 +48,7 @@ describe("GET API Route", () => {
         Body: stream,
       });
 
-    const response = await GET(request);
-    expect(response.status).toBe(200);
-    const jsonResponse = await response.json();
-    expect(jsonResponse.fhirBundle).toBeDefined();
+    const response = await getEcrData(fakeId);
+    expect(response).toBeDefined();
   });
 });
